fix: move components to the 'toolkit' folder
