+++ src/app/toolkit/layout.tsx
@@ -1,16 +1,14 @@
-
-'use client'
-import { ReactNode } from 'react';
-import Header from '../../ui/layout/ToolkitHeader';
+"use client";
+import { ReactNode } from "react";
+import Header from "./components/ToolkitHeader";
 
 export default function Page({ children }: { children: ReactNode }) {
   return (
     <div>
-     <div>
-      <Header />
+      <div>
+        <Header />
+      </div>
+      <div>{children}</div>
     </div>
-    <div>{children}</div>
-    </div>
-    
   );
 }
